♻️ Refactor: PhotoLog, PhotoLogs 관련 클래스 통합
