+++ core/network/src/main/java/com/twix/network/service/PhotoLogService.kt
@@ -1,7 +1,9 @@
 package com.twix.network.service
 
 import com.twix.network.model.response.photo.model.PhotoLogUploadUrlResponse
+import com.twix.network.model.response.photolog.PhotoLogResponse
 import de.jensklingenberg.ktorfit.http.GET
+import de.jensklingenberg.ktorfit.http.Path
 import de.jensklingenberg.ktorfit.http.Query
 
 interface PhotoLogService {
@@ -9,4 +11,9 @@ interface PhotoLogService {
     suspend fun getUploadUrl(
         @Query("goalId") goalId: Long,
     ): PhotoLogUploadUrlResponse
+
+    @GET("/api/v1/photologs/goals/{goalId}")
+    suspend fun fetchPhotoLogs(
+        @Path("goalId") goalId: Long,
+    ): PhotoLogResponse
 }
