+++ domain/src/main/java/com/twix/domain/repository/PhotoLogRepository.kt
@@ -1,6 +1,7 @@
 package com.twix.domain.repository
 
 import com.twix.domain.model.photo.PhotoLogUploadInfo
+import com.twix.domain.model.photolog.PhotoLogs
 import com.twix.result.AppResult
 
 interface PhotoLogRepository {
@@ -11,4 +12,6 @@ interface PhotoLogRepository {
         bytes: ByteArray,
         contentType: String,
     ): AppResult<String>
+
+    suspend fun fetchPhotoLogs(goalId: Long): AppResult<PhotoLogs>
 }
