--- domain/src/main/java/com/twix/domain/repository/PhotoLogsRepository.kt
@@ -1,8 +1,0 @@
-package com.twix.domain.repository
-
-import com.twix.domain.model.photolog.PhotoLogs
-import com.twix.result.AppResult
-
-interface PhotoLogsRepository {
-    suspend fun fetchPhotoLogs(goalId: Long): AppResult<PhotoLogs>
-}
